voice-assistant/requirements.txt
pipecat-ai==0.0.68
pipecat-ai[google]
pipecat-ai[local]
python-dotenv
loguru
websockets
firebase-admin
pywin32
pywinauto
selenium
webdriver_manager

# Google OAuth and API libraries
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
googleapis-common-protos


# Additional security and encryption
cryptography==41.0.7
keyring==24.3.0